# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: ketuk-api
  version: 0.2.0
  description: |
    Backend Ketuk. Sukses JSON `{"data": ...}` kecuali auth (`data` + `tokens`) dan callback Duitku (teks `SUCCESS`).
    Schema katalog + undangan + planner sesuai handler hidup. CRUD lokasi/tamu/buku ucapan sudah dikode sesuai `docs/contracts/undangan-crud.md`.
    PATCH inviter dan caption galeri juga sudah dikode sesuai `docs/contracts/undangan-patch-gallery-inviter.md`.
    CRUD planner (patch/archive, hari, kategori, item budget, checklist) sudah dikode sesuai `docs/contracts/planner-crud.md` bagian 1; rundown sesi jam+PIC belum.
    Auth dan billing masih path-only. Vendor belum diimplementasi.
servers:
  - url: http://localhost:8080
...
paths:
...
  /v1/planners:
    get:
      tags: [Planner]
      operationId: listPlanners
      summary: Daftar planner aktif
      description: "Hanya `status=active`. Planner yang sudah diarchive tetap terbaca lewat `GET /v1/planners/{id}`."
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlannerListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Planner]
      operationId: createPlanner
      summary: Buat planner (event | daily)
      description: "Kuota `max_active_planners`; mode `daily` butuh flag `planner_daily`. `categories` kosong + `type_id` → kategori dicopy dari `event_types.planner_categories`. Mata uang selalu IDR."
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlannerRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlannerResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          $ref: "#/components/responses/EntitlementDenied"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/planners/{id}:
    parameters:
      - $ref: "#/components/parameters/PlannerId"
    get:
      tags: [Planner]
      operationId: getPlanner
      summary: Detail planner
      description: Termasuk planner yang sudah diarchive, asal milik host.
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlannerResponse"
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      tags: [Planner]
      operationId: patchPlanner
      summary: Ubah planner (parsial)
      description: |
        Body parsial: field yang tidak dikirim tidak diubah, `{}` = no-op 200. `id` di body diabaikan.
        `type_id` / `starts_at` / `ends_at` dikirim `null` untuk dikosongkan.
        Turun `daily` → `event` mengosongkan `day_id` semua budget item; `day_id` checklist tetap.
        `status` dan `currency` tidak lewat PATCH — archive pakai DELETE, mata uang tetap IDR.
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchPlannerRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlannerResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          $ref: "#/components/responses/EntitlementDenied"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Planner]
      operationId: archivePlanner
      summary: Archive planner
      description: |
        Bukan hard delete: `status=archived`, baris budget dan checklist tetap di DB.
        Tautan `resource_links` tetap — hapus lewat `DELETE /v1/links/{id}`.
        Sudah archived → 204 idempoten. Membebaskan slot kuota `max_active_planners`. Unarchive belum ada.
      security: [{bearerAuth: []}]
      responses:
        "204":
          description: Diarchive
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/planners/{id}/days:
    parameters:
      - $ref: "#/components/parameters/PlannerId"
    get:
      tags: [Planner]
      operationId: listDays
      summary: Daftar hari
      description: "Urut `day_date`. Ubah / hapus satu hari: `/v1/planners/{id}/days/{dayId}`."
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlannerDayListResponse"
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Planner]
      operationId: addDay
      summary: Tambah hari
      description: "`date` = tanggal kalender (kolom DATE, jam diabaikan). Tanggal ganda dalam satu planner → 409 `conflict`."
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlannerDayRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlannerDayResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/DayDateConflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/planners/{id}/days/{dayId}:
    parameters:
      - $ref: "#/components/parameters/PlannerId"
      - $ref: "#/components/parameters/PlannerDayId"
    patch:
      tags: [Planner]
      operationId: patchDay
      summary: Ubah hari
      description: "Body parsial, `{}` = no-op 200. `label` boleh string kosong. Tanggal bentrok → 409 `conflict`. Hari bukan milik planner ini → 404 `day`."
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchPlannerDayRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlannerDayResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/DayDateConflict"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Planner]
      operationId: deleteDay
      summary: Hapus hari
      description: "Budget item dan checklist yang menempel hari ini tidak dihapus; `day_id` mereka jadi `null`."
      security: [{bearerAuth: []}]
      responses:
        "204":
          description: Dihapus
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/planners/{id}/days/{dayId}/sessions:
    parameters:
      - $ref: "#/components/parameters/PlannerId"
      - $ref: "#/components/parameters/PlannerDayId"
    get:
      tags: [Planner]
      operationId: listSessions
      summary: Daftar sesi rundown
      description: "Urut `starts_at` lalu `sort_order`. Hari bukan milik planner ini → 404 `day`."
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlannerSessionListResponse"
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Planner]
      operationId: addSession
      summary: Tambah sesi rundown
      description: "Jam dinding, bukan RFC3339: `\"HH:MM\"` atau `\"HH:MM:SS\"`. Tanggal ada di hari induknya. `pic_name` teks bebas, bukan id user/vendor."
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlannerSessionRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlannerSessionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/planners/{id}/days/{dayId}/sessions/{sessionId}:
    parameters:
      - $ref: "#/components/parameters/PlannerId"
      - $ref: "#/components/parameters/PlannerDayId"
      - $ref: "#/components/parameters/PlannerSessionId"
    patch:
      tags: [Planner]
      operationId: patchSession
      summary: Ubah sesi rundown
      description: "Body parsial, `{}` = no-op 200. `ends_at` menerima `null` untuk sesi tanpa jam selesai. Sesi di hari lain → 404 `session`."
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchPlannerSessionRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlannerSessionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Planner]
      operationId: deleteSession
      summary: Hapus sesi rundown
      description: "Hapus hari ikut menghapus sesinya (FK CASCADE)."
      security: [{bearerAuth: []}]
      responses:
        "204":
          description: Dihapus
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/planners/{id}/categories:
    parameters:
      - $ref: "#/components/parameters/PlannerId"
    get:
      tags: [Planner]
      operationId: listCategories
      summary: Daftar kategori budget
      description: "Ubah / hapus satu kategori: `/v1/planners/{id}/categories/{categoryId}`."
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetCategoryListResponse"
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Planner]
      operationId: addCategory
      summary: Tambah kategori budget
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBudgetCategoryRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetCategoryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/planners/{id}/categories/{categoryId}:
    parameters:
      - $ref: "#/components/parameters/PlannerId"
      - $ref: "#/components/parameters/BudgetCategoryId"
    patch:
      tags: [Planner]
      operationId: patchCategory
      summary: Ubah kategori budget
      description: "Body parsial, `{}` = no-op 200. `name` kosong → 400. Kategori bukan milik planner ini → 404 `category`."
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchBudgetCategoryRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetCategoryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Planner]
      operationId: deleteCategory
      summary: Hapus kategori budget
      description: "**Cascade:** semua budget item di dalam kategori ini terhapus. Minta konfirmasi host sebelum memanggil."
      security: [{bearerAuth: []}]
      responses:
        "204":
          description: Dihapus bersama itemnya
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/planners/{id}/budget/items:
    parameters:
      - $ref: "#/components/parameters/PlannerId"
    post:
      tags: [Planner]
      operationId: addBudgetItem
      summary: Tambah item budget
      description: |
        `category_id` harus kategori milik planner ini (salah → 404 `category`). `amount` IDR, `< 0` → 400.
        `day_id` hanya boleh saat `budget_mode` = `daily` (mode `event` → 400); hari asing → 404 `day`.
        Baca item lewat `GET /v1/planners/{id}/budget`.
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBudgetItemRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetItemResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/planners/{id}/budget/items/{itemId}:
    parameters:
      - $ref: "#/components/parameters/PlannerId"
      - $ref: "#/components/parameters/BudgetItemId"
    patch:
      tags: [Planner]
      operationId: patchBudgetItem
      summary: Ubah item budget
      description: |
        Body parsial, `{}` = no-op 200. Item bukan milik planner ini → 404 `budget_item`.
        `paid_at` RFC3339 = sudah dibayar, `null` = belum. Server tidak mengisi `now()` sendiri.
        `day_id` `null` = lepas dari hari; non-null hanya saat mode `daily`.
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchBudgetItemRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetItemResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Planner]
      operationId: deleteBudgetItem
      summary: Hapus item budget
      security: [{bearerAuth: []}]
      responses:
        "204":
          description: Dihapus
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
...
    PlannerId:
      name: id
      in: path
      required: true
      schema: {type: string, format: uuid}
    PlannerDayId:
      name: dayId
      in: path
      required: true
      schema: {type: string, format: uuid}
    PlannerSessionId:
      name: sessionId
      in: path
      required: true
      schema: {type: string, format: uuid}
...
  responses:
    BadRequest:
      description: "JSON rusak (`invalid_json`) atau validasi (`invalid_input` / `invalid_id`)"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    InvalidId:
      description: Path UUID rusak
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
          example:
            error: {code: invalid_id, message: invalid id}
    Unauthorized:
      description: JWT hilang atau tidak valid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
          example:
            error: {code: unauthorized, message: missing bearer token}
    EntitlementDenied:
      description: Kuota atau template premium
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
          examples:
            invitationQuota:
              value:
                error: {code: entitlement_denied, message: max_active_invitations quota exceeded}
            guestQuota:
              value:
                error: {code: entitlement_denied, message: max_guests_per_invitation quota exceeded}
            premiumTemplate:
              value:
                error: {code: entitlement_denied, message: premium_templates not included in current plan}
            plannerQuota:
              value:
                error: {code: entitlement_denied, message: max_active_planners quota exceeded}
            plannerDaily:
              value:
                error: {code: entitlement_denied, message: planner_daily not included in current plan}
    NotFound:
      description: Resource tidak ada atau bukan milik host; RSVP token/slug unpublished
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
          examples:
            invitation:
              value:
                error: {code: not_found, message: invitation not found}
            plannerChild:
              summary: "Anak planner: day / session / category / budget_item / checklist"
              value:
                error: {code: not_found, message: budget_item not found}
    DayDateConflict:
      description: Tanggal hari sudah ada di planner ini
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
          example:
            error: {code: conflict, message: day date already exists}
...
    InternalError:
      description: Kesalahan server
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
          example:
            error: {code: internal, message: internal server error}
  schemas:
...
    CreatePlannerRequest:
      type: object
      required: [title]
      properties:
        title: {type: string}
        budget_mode: {type: string, enum: [event, daily], default: event}
        type_id: {type: string, format: uuid, nullable: true}
        categories:
          type: array
          nullable: true
          items: {type: string}
          description: Nama kategori awal; kosong + `type_id` → dicopy dari `event_types.planner_categories`
    PatchPlannerRequest:
      type: object
      description: "Field yang tidak dikirim tidak diubah. `type_id` / `starts_at` / `ends_at` menerima `null` untuk dikosongkan."
      properties:
        title: {type: string, description: Trim; kosong → 400}
        budget_mode: {type: string, enum: [event, daily]}
        type_id: {type: string, format: uuid, nullable: true}
        starts_at: {type: string, format: date-time, nullable: true}
        ends_at: {type: string, format: date-time, nullable: true, description: Tanggal sebelum `starts_at` → 400}
...
    CreatePlannerDayRequest:
      type: object
      required: [date]
      properties:
        date: {type: string, format: date-time}
        label: {type: string}
    PatchPlannerDayRequest:
      type: object
      properties:
        date: {type: string, format: date-time}
        label: {type: string, description: String kosong diperbolehkan}
        sort_order: {type: integer}
...
    PlannerResponse:
      type: object
      required: [data]
      properties:
        data: {$ref: "#/components/schemas/Planner"}
    PlannerListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          nullable: true
          items: {$ref: "#/components/schemas/Planner"}
    PlannerDayResponse:
      type: object
      required: [data]
      properties:
        data: {$ref: "#/components/schemas/PlannerDay"}
    PlannerDayListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          nullable: true
          items: {$ref: "#/components/schemas/PlannerDay"}
    PlannerSessionResponse:
      type: object
      required: [data]
      properties:
        data: {$ref: "#/components/schemas/PlannerSession"}
    PlannerSessionListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          nullable: true
          items: {$ref: "#/components/schemas/PlannerSession"}
...